Refactor remote method completion context
