+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/completions/providers/context/RemoteMethodCallActionNodeContext.java
@@ -15,13 +15,20 @@
  */
 package org.ballerinalang.langserver.completions.providers.context;
 
+import io.ballerina.compiler.api.symbols.Symbol;
+import io.ballerina.compiler.api.symbols.TypeSymbol;
 import io.ballerina.compiler.syntax.tree.RemoteMethodCallActionNode;
 import org.ballerinalang.annotation.JavaSPIService;
+import org.ballerinalang.langserver.common.utils.SymbolUtil;
 import org.ballerinalang.langserver.commons.BallerinaCompletionContext;
 import org.ballerinalang.langserver.commons.completion.LSCompletionException;
 import org.ballerinalang.langserver.commons.completion.LSCompletionItem;
+import org.ballerinalang.langserver.completions.util.ContextTypeResolver;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Handles the completions for the {@link RemoteMethodCallActionNode}.
@@ -38,9 +45,39 @@ public class RemoteMethodCallActionNodeContext extends RightArrowActionNodeConte
     @Override
     public List<LSCompletionItem> getCompletions(BallerinaCompletionContext context, RemoteMethodCallActionNode node)
             throws LSCompletionException {
-        List<LSCompletionItem> completionItems = this.getFilteredItems(context, node, node.expression());
+        List<LSCompletionItem> completionItems = new ArrayList<>();
+        ContextTypeResolver resolver = new ContextTypeResolver(context);
+        Optional<TypeSymbol> expressionType = node.expression().apply(resolver);
+
+        if (expressionType.isEmpty() || !SymbolUtil.isClient(expressionType.get())) {
+            return Collections.emptyList();
+        }
+
+        if (onSuggestClientActions(node, context)) {
+            /*
+            Covers the following case where a is a client object and we suggest the remote actions
+            (1) a -> g<cursor>
+             */
+            List<Symbol> clientActions = this.getClientActions(expressionType.get());
+            completionItems.addAll(this.getCompletionItemList(clientActions, context));
+        } else if (onSuggestFunctionParameters(node, context)) {
+            completionItems.addAll(this.expressionCompletions(context));
+        }
+
         this.sort(context, node, completionItems);
 
         return completionItems;
+    }
+
+    private boolean onSuggestClientActions(RemoteMethodCallActionNode node, BallerinaCompletionContext context) {
+        int cursor = context.getCursorPositionInTree();
+        return node.rightArrowToken().textRange().endOffset() <= cursor &&
+                (node.openParenToken().isMissing() || cursor <= node.openParenToken().textRange().startOffset());
+    }
+
+    private boolean onSuggestFunctionParameters(RemoteMethodCallActionNode node, BallerinaCompletionContext context) {
+        int cursor = context.getCursorPositionInTree();
+        return !node.openParenToken().isMissing() && node.openParenToken().textRange().endOffset() <= cursor &&
+                (node.closeParenToken().isMissing() || cursor <= node.closeParenToken().textRange().startOffset());
     }
 }
